requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
opencv-python
Pillow
tqdm
gradio
datasets
kaggle
torchserve
torch-model-archiver
prometheus-client
pyyaml
matplotlib
albumentations
